Final Submission Has Been Accepted V1

+++ src/Applications/use_case/comments/_test/DeleteCommentUseCase.test.js
@@ -1,28 +1,49 @@
 const CommentRepository = require('../../../../Domains/comments/CommentRepository');
-const ThreadRepository = require('../../../../Domains/threads/ThreadRepository');
 const DeleteCommentUseCase = require('../DeleteCommentUseCase');
 
 describe('DeleteCommentUseCase', () => {
     it('should orchestracting the delete comment action correctly', async () => {
         // Arrange
-        const mockCommentRepository = new CommentRepository();
-        const mockThreadRepository = new ThreadRepository();
+        const useCasePayload = {
+            threadId: 'thread-123',
+            commentId: 'comment-123',
+            userId: 'user-123',
+        };
 
+        /** creating dependency of use case */
+        const mockCommentRepository = new CommentRepository();
+
+        /** mocking needed function */
         mockCommentRepository.deleteCommentById = jest.fn()
             .mockImplementation(() => Promise.resolve());
-        mockThreadRepository.verifyThread = jest.fn()
+        mockCommentRepository.verifyCommentAccess = jest.fn()
+            .mockImplementation(() => Promise.resolve());
+        mockCommentRepository.verifyCommentIsExist = jest.fn()
             .mockImplementation(() => Promise.resolve());
 
-        const usecase = new DeleteCommentUseCase({
+        /** creating use case instance */
+        const deleteCommentUseCase = new DeleteCommentUseCase({
             commentRepository: mockCommentRepository,
-            threadRepository: mockThreadRepository,
         });
 
         // Action
-        await usecase.execute('thread-123', 'comment-123', 'user-123');
+        await deleteCommentUseCase.execute({
+            threadId: useCasePayload.threadId,
+            commentId: useCasePayload.commentId,
+            userId: useCasePayload.userId,
+        });
 
         // Assert
-        expect(mockThreadRepository.verifyThread).toBeCalledWith('thread-123');
-        expect(mockCommentRepository.deleteCommentById).toBeCalledWith('comment-123', 'user-123');
+        expect(mockCommentRepository.verifyCommentIsExist).toBeCalledWith(
+            useCasePayload.commentId,
+            useCasePayload.threadId,
+        );
+        expect(mockCommentRepository.verifyCommentAccess).toBeCalledWith(
+            useCasePayload.commentId,
+            useCasePayload.userId,
+        );
+        expect(mockCommentRepository.deleteCommentById).toBeCalledWith(
+            useCasePayload.commentId,
+        );
     });
 });
