+++ src/Applications/use_case/replies/_test/DeleteReplyUseCase.test.js
@@ -1,34 +1,43 @@
-const ThreadRepository = require('../../../../Domains/threads/ThreadRepository');
-const CommentRepository = require('../../../../Domains/comments/CommentRepository');
 const RepliesRepository = require('../../../../Domains/replies/RepliesRepository');
 const DeleteReplyUseCase = require('../DeleteReplyUseCase');
 
 describe('DeleteReplyUseCase', () => {
     it('should orchestracting the delete reply action correctly', async () => {
         // Arrange
-        const mockRepliesRepository = new RepliesRepository();
-        const mockCommentRepository = new CommentRepository();
-        const mockThreadRepository = new ThreadRepository();
+        const useCaseDeletePayload = {
+            threadId: 'thread-123',
+            commentId: 'comment-123',
+            replyId: 'reply-123',
+            userId: 'user-123',
+        };
 
+        /** creating dependency of use case */
+        const mockRepliesRepository = new RepliesRepository();
+
+        /** mocking needed function */
         mockRepliesRepository.deleteReply = jest.fn()
             .mockImplementation(() => Promise.resolve());
-        mockCommentRepository.verifyCommentLocation = jest.fn()
+        mockRepliesRepository.verifyReplyAccess = jest.fn()
             .mockImplementation(() => Promise.resolve());
-        mockThreadRepository.verifyThread = jest.fn()
+        mockRepliesRepository.verifyReplyIsExist = jest.fn()
             .mockImplementation(() => Promise.resolve());
 
-        const usecase = new DeleteReplyUseCase({
+        /** creating use case instance */
+        const deleteReplyUseCase = new DeleteReplyUseCase({
             repliesRepository: mockRepliesRepository,
-            commentRepository: mockCommentRepository,
-            threadRepository: mockThreadRepository,
         });
 
         // Action
-        await usecase.execute('thread-123', 'comment-123', 'reply-123', 'user-123');
+        await deleteReplyUseCase.execute(useCaseDeletePayload);
 
         // Assert
-        expect(mockThreadRepository.verifyThread).toBeCalledWith('thread-123');
-        expect(mockCommentRepository.verifyCommentLocation).toBeCalledWith('comment-123', 'thread-123');
-        expect(mockRepliesRepository.deleteReply).toBeCalledWith('reply-123', 'user-123');
+        expect(mockRepliesRepository.verifyReplyIsExist).toBeCalledWith(
+            useCaseDeletePayload.replyId,
+        );
+        expect(mockRepliesRepository.verifyReplyAccess).toBeCalledWith(
+            useCaseDeletePayload.replyId,
+            useCaseDeletePayload.userId,
+        );
+        expect(mockRepliesRepository.deleteReply).toBeCalledWith(useCaseDeletePayload.replyId);
     });
 });
